backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==4.3.0
black==25.1.0
boto3==1.34.14
botocore==1.34.162
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
environs==14.3.0
fastapi==0.109.0
fastapi_cors==0.0.6
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.1
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.2
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
